fix: net-firewall/iptables test failures
The iptables test suite encounters multiple failures on the 6.1.67
Gentoo distribution kernel. This enables the nftables and conntrack
USE flags on iptables and the xtables USE flag on nftables to fix most
of the failures. The remaining failures in xlate-test.py are fixed by
reverting the upsteam commit 5839d7f. That causes the ip6t_mh tests to
expect the protocol number 135 instead of the name "mobility-header".

Refs: https://bugs.gentoo.org/890628
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>