fix: dev-cpp/catch test failures
The dev-cpp/catch tests fail because the paths in the expected test
output differ from the paths in the actual test output. This is because
they are prefixes with "Catch2-3.4.0/" in the actual output. This adds
a patch to remove "Catch2-<version>/" in detected paths. The patch also
updates the re modules usage so regex strings are properly prefixed with
"r".

Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>