fix: dev-libs/check test failure
The dev-libs/check test suite is failing because the expected output
is not matching the actual output. This is because the working directory
the tests are run from is producing absolute paths when relative paths
are expected. This patch fixes the issue by setting the SRCDIR variable
explicitly to a relative path using realpath.

Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>