fix: dev-python/cython test failures
The cython tests are failing for multiple reasons which are fixed by the
two patches added in this commit.

The first patch reverts a change made to pyximport to better support
windows which was causing a test failure when run in the ebuild.

The second patch fixes a test failure caused by the test not being aware
of how Gentoo handles pip packages. It adds the --break-system-packages
flag to work around the test failure which doesn't seem to actually
break any system packages.

Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>