fix: sys-libs/glibc test failure
glibc tests its ability to detect CPU features by comparing its results
to the /proc/cpuinfo file. The Intel TSX vulnerability called ZombieLoad
2 caused Intel to release a microcode update that disables TSX on
impacted CPUs. Disability TSX has the effect of removing the Hardware
Lock Elision (HLE) CPU flag as well.

The initial microcode update (IPU 2021.1) was updated in IPU 2021.2 to
disable TSX on additional CPUs. glibc has not been updated to account
for this change. This patch adds the new CPU steppings to the list of
steppings that should not support HLE. This issue was encountered on a
Dell Precision 7540 with an Intel Core i9-9980HK CPU.

Refs: https://www.intel.com/content/www/us/en/support/articles/000059422/processors.html
Signed-off-by: Jaremy Hatler <hatler.jaremy@gmail.com>